Hide scroll indicator
